+++ include/polyscope/scalar_render_image_quantity.h
@@ -17,7 +17,7 @@ class ScalarRenderImageQuantity : public RenderImageQuantityBase, public ScalarQ
 public:
   ScalarRenderImageQuantity(Structure& parent_, std::string name, size_t dimX, size_t dimY,
                             const std::vector<float>& depthData, const std::vector<glm::vec3>& normalData,
-                            const std::vector<double>& scalarData, ImageOrigin imageOrigin, DataType dataType);
+                            const std::vector<float>& scalarData, ImageOrigin imageOrigin, DataType dataType);
 
   virtual void draw() override;
   virtual void drawDelayed() override;
@@ -52,7 +52,7 @@ void ScalarRenderImageQuantity::updateBuffers(const T1& depthData, const T2& nor
   // standardize
   std::vector<float> standardDepth(standardizeArray<float>(depthData));
   std::vector<glm::vec3> standardNormal(standardizeVectorArray<glm::vec3, 3>(normalData));
-  std::vector<double> standardScalar(standardizeArray<double>(scalarData));
+  std::vector<float> standardScalar(standardizeArray<float>(scalarData));
 
   values.data = standardScalar;
   values.markHostBufferUpdated();
